requirements.txt
minerl==0.3.7
ray
gym
pyyaml
numpy
